Updated rewards
Updated rewards for more free learning (less bias)
Added winning condition and logging in the Play.py for experiments sake :D
Chnaged some other minor stuff and run a lot of experiments as to fine tune the rewards and the scale. in the end quite good results.

Out of 310, 258 met the landing condition and 42 just ran out of time ;D