Added Avular Config and Demo File
Added the new Avular USD to the Project as we're switching from the craxyflie to the real drone.
I also added a Demo File that when played directly from Visual Studio Code, it loads the Drone in. (Currently not working on the machine im using as it's trowing an Isaac lab import not known but I'll try it on my machine at home