Working MARL agent and use of checkpoints
The MARL agent is now live.
It can be trained and played. At the moment, it currently only supports gusts and not constant wind. (to be implemented)

The tracking is not yet perfect but works for the success and magnet (not for crash and timeout)

I am now able to use checkpoints and train from them.
I have equally also updated the read me file with the checkpoint mechanic.